Added get company by discount id feature

+++ src/main/java/com/dealuni/demo/controllers/UserController.java
@@ -2,9 +2,13 @@ package com.dealuni.demo.controllers;
 
 import com.dealuni.demo.dto.UserRequest;
 import com.dealuni.demo.dto.UserResponse;
+import com.dealuni.demo.models.CustomUserDetails;
 import com.dealuni.demo.services.UserService;
 import jakarta.validation.Valid;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -23,6 +27,7 @@ public class UserController {
     }
 
     //get all users
+    @PreAuthorize("hasRole('ADMIN')")
     @GetMapping
     public ResponseEntity<List<UserResponse>> getAllUsers() {
         List<UserResponse> userResponseList = userService.getAllExistingUsers();
@@ -38,15 +43,32 @@ public class UserController {
 
     //update user by id
     @PatchMapping("/{id}")
-    public ResponseEntity<UserResponse> updateUser(@PathVariable Long id, @Valid @RequestBody UserRequest userRequest) {
+    public ResponseEntity<UserResponse> updateUser(@PathVariable Long id, @Valid @RequestBody UserRequest userRequest, @AuthenticationPrincipal CustomUserDetails userDetails) {
+
+        // dacă userul NU e admin și vrea să editeze pe altcineva -> interzis
+        boolean isAdmin = userDetails.getAuthorities().stream().anyMatch(auth -> auth.getAuthority().equals("ROLE_ADMIN"));
+
+        if (!isAdmin && !userDetails.getId().equals(id)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         UserResponse userResponse = userService.updateUserById(id, userRequest);
         return ResponseEntity.ok(userResponse);
     }
 
+
     //delete user by id
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteUserById(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteUserById(@PathVariable Long id, @AuthenticationPrincipal CustomUserDetails userDetails) {
+
+        boolean isAdmin = userDetails.getAuthorities().stream().anyMatch(auth -> auth.getAuthority().equals("ROLE_ADMIN"));
+
+        if (!isAdmin && !userDetails.getId().equals(id)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         userService.deleteUserById(id);
         return ResponseEntity.noContent().build();
     }
+
 }
